LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight-star.any-expected.txt

CONSOLE MESSAGE: Method OK is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=*&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=*&allow_headers=& due to access control checks.
CONSOLE MESSAGE: Request header field X-Test is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=get&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Request header field X-Test is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=*&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method DELETE is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=delete&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method DELETE is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=delete&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=put&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=put&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method patch is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=PATCH&allow_headers=*& due to access control checks.
CONSOLE MESSAGE: Method PATCH is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/preflight.py?origin=http://web-platform.test:8800&credentials&allow_methods=patch&allow_headers=*& due to access control checks.

PASS CORS that succeeds with credentials: false; method: GET (allowed: get); header: X-Test,1 (allowed: x-test)
PASS CORS that succeeds with credentials: false; method: SUPER (allowed: *); header: X-Test,1 (allowed: x-test)
PASS CORS that succeeds with credentials: false; method: OK (allowed: *); header: X-Test,1 (allowed: *)
PASS CORS that fails with credentials: true; method: OK (allowed: *); header: X-Test,1 (allowed: *)
PASS CORS that fails with credentials: true; method: PUT (allowed: *); header:  (allowed: )
PASS CORS that fails with credentials: true; method: GET (allowed: get); header: X-Test,1 (allowed: *)
PASS CORS that fails with credentials: true; method: GET (allowed: *); header: X-Test,1 (allowed: *)
PASS CORS that succeeds with credentials: true; method: * (allowed: *); header: *,1 (allowed: *)
PASS CORS that succeeds with credentials: true; method: GET (allowed: GET); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: get (allowed: GET); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: GET (allowed: get); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: get (allowed: get); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: HEAD (allowed: HEAD); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: head (allowed: HEAD); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: HEAD (allowed: head); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: head (allowed: head); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: POST (allowed: POST); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: post (allowed: POST); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: POST (allowed: post); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: post (allowed: post); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: DELETE (allowed: DELETE); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: delete (allowed: DELETE); header:  (allowed: *)
PASS CORS that fails with credentials: true; method: DELETE (allowed: delete); header:  (allowed: *)
PASS CORS that fails with credentials: true; method: delete (allowed: delete); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: PUT (allowed: PUT); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: put (allowed: PUT); header:  (allowed: *)
PASS CORS that fails with credentials: true; method: PUT (allowed: put); header:  (allowed: *)
PASS CORS that fails with credentials: true; method: put (allowed: put); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: PATCH (allowed: PATCH); header:  (allowed: *)
PASS CORS that fails with credentials: true; method: patch (allowed: PATCH); header:  (allowed: *)
PASS CORS that fails with credentials: true; method: PATCH (allowed: patch); header:  (allowed: *)
PASS CORS that succeeds with credentials: true; method: patch (allowed: patch); header:  (allowed: *)
